added name my trip input to home page seprate modal to another file from home page
